test(node): add thread safety test
[ci skip] no need to run on ci

+++ src/cpp/svn_error.hpp
@@ -6,19 +6,19 @@
 namespace svn {
 class svn_error : public std::runtime_error {
   public:
-    explicit svn_error(int         code,
+    explicit svn_error(int32_t     code,
                        const char* what,
                        svn_error*  child = nullptr,
                        std::string file  = std::string(),
-                       int         line  = -1) noexcept;
+                       int32_t     line  = -1) noexcept;
     svn_error(const svn_error& other) noexcept;
     svn_error(svn_error&& other) noexcept;
 
     ~svn_error();
 
-    int         code;
+    int32_t     code;
     svn_error*  child;
     std::string file;
-    long        line;
+    int32_t     line;
 };
 } // namespace svn
